Improve cipher API to reduce chance of nonce reuse
This commit changes the encryption functions of `asymmetric::cipher` and
`symmetric::cipher` to accept an `Option<&Nonce>` for the nonce
parameter, rather than just a plain `Nonce`. If `None` is specified for
this argument, then alkali will now randomly generate a nonce, rather
than relying on the user to prevent nonce reuse.