Allow protected memory to be shared across threads
The `hardend_buffer` macro generates all keys and other important data that need to be protected. The underlying struct is based on a pointer, hence by default the `Send` and `Sync` marker traits are not implemented, making it impossible to share such important data between threads.

Given that each new type have exclusive access to its pointer, and given that no mutability is possible from a read-only reference, and for the same reasons why it is safe to `Deref` and `DerefMut` to a `[u8; $size]`, it is safe to share such a memory.