Improve kdf::derive_key type-safety
When we want to derive a new key with `hash::kdf::derive_key` we need to
pass context as a `&str`, however it has a strict requirement on length
that is checked at runtime.

We can move this error at compile-time by changing the type to
`&[u8; CONTEXT_SIZE]`. Using this type in rust is quite convenient,
instead of `"UserName"` we now have `b"UserName"`.

Another requirement is that context must not contain any null byte
`0x00`, however passing a string such as `"UserNam\0"` before would
panic. We now surface this error to the user, which is less surprising
as a behaviour.