Switch from type definitions to tuple structs
This commit switches a number of array type aliases to actually be tuple
structs wrapping an array of the same length. This is done to improve
Serde support - As const generics have not yet landed in Serde,
Serialize & Deserialize are only automatically implemented for arrays
with up to 32 elements. For larger arrays, a manual implementation is
required.

We wrap long arrays in a struct with an obvious Serialize/Deserialize
implementation to improve the developer experience when using types like
`Signature`.

Fixes #1