Add Mprotect generic param to hardened buffers
This change lays the groundwork for support for the `sodium_mprotect`
API. The intention is to provide a type-safe system for marking hardened
buffer types as readonly/no-access. To this end, we introduce a new
generic parameter on all hardened buffer types, which must implement
`alkali::mem::MprotectStatus`. This is a parameter used as a marker,
indicating the current mprotect status of the type, which can be one of
`alkali::mem::FullAccess`, `alkali::mem::ReadOnly`, or
`alkali::mem::NoAccess`. There is also a trait
`alkali::mem::MprotectReadable` implemented for `FullAccess` and
`ReadOnly` which can be used wherever a hardened buffer need only be
read.

We now only implement methods which will mutate the hardened buffer if
the `Mprotect` parameter is `FullAccess`, and methods which will read
from the hardened buffer if the `Mprotect` parameter is
`MprotectReadable`.

In the near future, methods for converting between these mprotect
statuses will be introduced (and we'll actually start protecting memory,
rather than just introducing the machinery which makes it possible to do
so).