Update example to align with best practices
This commit updates key management throughout the file encryption
example. First, the key is protected as read-only immediately after
derivation. Second, as soon as the key is no longer required, it is
freed (Sodium's allocator will zero the memory before free, to avoid
recovery from uninitialised memory).

Error handling for the example has also been updated, although it is
still subpar in comparison to production-quality code, since this is
just a demo.